kafka_api_consumer/requirements.txt
requests
confluent-kafka==2.8.0
python-dotenv
azure-storage-blob==12.24.0
boto3==1.35.73
pandas==2.2.3
pyarrow==19.0.1
s3fs==2025.2.0
colorlog==6.9.0
fsspec==2025.2.0
